Solution to Check Positive or Negative

+++ Check Positive or Negative@@6788dd63490637cfc44e1b8f/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main(){
-    int a;
-    scanf("%d",&a);
+    float a;
+    scanf("%f",&a);
     if(a>0){
         printf("Positive");
     }
